+++ NEWS.md
@@ -29,6 +29,7 @@
 ## Coercion
 
 - Improved `as_stocnet.igraph()` so that it maps a 'lvl' attribute onto the 'mode' variable of the nodes table, naming the levels from the network's info where there is a name for each
+- Fixed `as_igraph()` erroring on an unlabelled stocnet object with nodal attributes, e.g. `as_igraph(as_stocnet(create_ring(8)))`
 
 ## Manipulating